+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - By default implies `-q / --hide-control-chars` if the stdout is connected to a terminal, otherwise `--show-control-chars`.
 - No surrond a divition by any SGR parameter when no appropriate one found.
 - The default `TIME_STYLE` is `$'%b %e  %Y\n%b %e %H:%M'`.
+- Show `?` as missing file's information with long format.
 
 ### Fixed